Splitted PopupAnnotations.
