+++ ts/Extensions/Annotations/Popup/Popup.ts
@@ -25,6 +25,7 @@ import NavigationBindings from '../NavigationBindings.js';
 import D from '../../../Core/DefaultOptions.js';
 const { getOptions } = D;
 import Pointer from '../../../Core/Pointer.js';
+import PopupAnnotations from './PopupAnnotations.js';
 import PopupIndicators from './PopupIndicators.js';
 import PopupTabs from './PopupTabs.js';
 import U from '../../../Core/Utilities.js';
@@ -33,11 +34,7 @@ const {
     createElement,
     extend,
     fireEvent,
-    isArray,
-    isObject,
-    objectEach,
     pick,
-    stableSort,
     wrap
 } = U;
 
@@ -50,30 +47,6 @@ declare global {
         interface IndicatorNameCouple {
             indicatorFullName: string;
             indicatorType: string;
-        }
-        interface PopupAnnotationsObject {
-            addForm(
-                this: Popup,
-                chart: AnnotationChart,
-                options: AnnotationsOptions,
-                callback: Function,
-                isInit?: boolean
-            ): void;
-            addFormFields(
-                this: Popup,
-                parentDiv: HTMLDOMElement,
-                chart: AnnotationChart,
-                parentNode: string,
-                options: AnnotationsOptions,
-                storage: Array<unknown>,
-                isRoot?: boolean
-            ): void;
-            addToolbar(
-                this: Popup,
-                chart: AnnotationChart,
-                options: AnnotationsOptions,
-                callback: Function
-            ): void;
         }
         interface PopupConfigObject {
             annotation: Annotation;
@@ -102,278 +75,6 @@ declare global {
         }
     }
 }
-
-
-const annotations: Highcharts.PopupAnnotationsObject = {
-    /**
-     * Create annotation simple form. It contains two buttons
-     * (edit / remove) and text label.
-     * @private
-     * @param {Highcharts.Chart} - chart
-     * @param {Highcharts.AnnotationsOptions} - options
-     * @param {Function} - on click callback
-     */
-    addToolbar: function (
-        this: Popup,
-        chart: Highcharts.AnnotationChart,
-        options: AnnotationsOptions,
-        callback: Function
-    ): void {
-        let _self = this,
-            lang = this.lang,
-            popupDiv = this.container,
-            showForm = this.showForm,
-            toolbarClass = PREFIX + 'annotation-toolbar',
-            button;
-
-        // set small size
-        if (popupDiv.className.indexOf(toolbarClass) === -1) {
-            popupDiv.className += ' ' + toolbarClass;
-        }
-
-        // set position
-        if (chart) {
-            popupDiv.style.top = chart.plotTop + 10 + 'px';
-        }
-
-        // create label
-        createElement(SPAN, void 0, void 0, popupDiv).appendChild(
-            doc.createTextNode(pick(
-                // Advanced annotations:
-                lang[options.langKey as any] || options.langKey,
-                // Basic shapes:
-                options.shapes && options.shapes[0].type,
-                ''
-            ))
-        );
-
-        // add buttons
-        button = this.addButton(
-            popupDiv,
-            lang.removeButton || 'remove',
-            'remove',
-            popupDiv,
-            callback
-        );
-
-        button.className += ' ' + PREFIX + 'annotation-remove-button';
-        button.style['background-image' as any] = 'url(' +
-            this.iconsURL + 'destroy.svg)';
-
-        button = this.addButton(
-            popupDiv,
-            lang.editButton || 'edit',
-            'edit',
-            popupDiv,
-            function (): void {
-                showForm.call(
-                    _self,
-                    'annotation-edit',
-                    chart,
-                    options,
-                    callback
-                );
-            }
-        );
-
-        button.className += ' ' + PREFIX + 'annotation-edit-button';
-        button.style['background-image' as any] = 'url(' +
-            this.iconsURL + 'edit.svg)';
-
-    },
-    /**
-     * Create annotation simple form.
-     * It contains fields with param names.
-     * @private
-     * @param {Highcharts.Chart} chart
-     * Chart
-     * @param {Object} options
-     * Options
-     * @param {Function} callback
-     * On click callback
-     * @param {boolean} [isInit]
-     * If it is a form declared for init annotation
-     */
-    addForm: function (
-        this: Popup,
-        chart: Highcharts.AnnotationChart,
-        options: AnnotationsOptions,
-        callback: Function,
-        isInit?: boolean
-    ): void {
-        let popupDiv = this.container,
-            lang = this.lang,
-            bottomRow,
-            lhsCol;
-
-        if (!chart) {
-            return;
-        }
-
-        // create title of annotations
-        lhsCol = createElement('h2', {
-            className: PREFIX + 'popup-main-title'
-        }, void 0, popupDiv);
-        lhsCol.appendChild(
-            doc.createTextNode(
-                lang[options.langKey as any] || options.langKey || ''
-            )
-        );
-
-        // left column
-        lhsCol = createElement(
-            DIV,
-            {
-                className: (
-                    PREFIX + 'popup-lhs-col ' + PREFIX + 'popup-lhs-full'
-                )
-            },
-            void 0,
-            popupDiv
-        );
-
-        bottomRow = createElement(
-            DIV,
-            {
-                className: PREFIX + 'popup-bottom-row'
-            },
-            void 0,
-            popupDiv
-        );
-
-        this.annotations.addFormFields.call(
-            this,
-            lhsCol,
-            chart,
-            '',
-            options,
-            [],
-            true
-        );
-
-        this.addButton(
-            bottomRow,
-            isInit ?
-                (lang.addButton || 'add') :
-                (lang.saveButton || 'save'),
-            isInit ? 'add' : 'save',
-            popupDiv,
-            callback
-        );
-    },
-    /**
-     * Create annotation's form fields.
-     * @private
-     * @param {Highcharts.HTMLDOMElement} parentDiv
-     * Div where inputs are placed
-     * @param {Highcharts.Chart} chart
-     * Chart
-     * @param {string} parentNode
-     * Name of parent to create chain of names
-     * @param {Highcharts.AnnotationsOptions} options
-     * Options
-     * @param {Array<unknown>} storage
-     * Array where all items are stored
-     * @param {boolean} [isRoot]
-     * Recursive flag for root
-     */
-    addFormFields: function (
-        this: Popup,
-        parentDiv: HTMLDOMElement,
-        chart: Highcharts.AnnotationChart,
-        parentNode: string,
-        options: AnnotationsOptions,
-        storage: Array<unknown>,
-        isRoot?: boolean
-    ): void {
-        let _self = this,
-            addFormFields = this.annotations.addFormFields,
-            addInput = this.addInput,
-            lang = this.lang,
-            parentFullName,
-            titleName;
-
-        if (!chart) {
-            return;
-        }
-
-        objectEach(options, function (value, option: string): void {
-
-            // create name like params.styles.fontSize
-            parentFullName = parentNode !== '' ?
-                parentNode + '.' + option : option;
-
-            if (isObject(value)) {
-                if (
-                    // value is object of options
-                    !isArray(value) ||
-                    // array of objects with params. i.e labels in Fibonacci
-                    (isArray(value) && isObject(value[0]))
-                ) {
-                    titleName = lang[option] || option;
-
-                    if (!titleName.match(indexFilter)) {
-                        storage.push([
-                            true,
-                            titleName,
-                            parentDiv
-                        ]);
-                    }
-
-                    addFormFields.call(
-                        _self,
-                        parentDiv,
-                        chart,
-                        parentFullName,
-                        value as any,
-                        storage,
-                        false
-                    );
-                } else {
-                    storage.push([
-                        _self,
-                        parentFullName,
-                        'annotation',
-                        parentDiv,
-                        value
-                    ]);
-                }
-            }
-        });
-
-        if (isRoot) {
-            stableSort(storage, function (a: any): number {
-                return a[1].match(/format/g) ? -1 : 1;
-            });
-
-            if (isFirefox) {
-                storage.reverse(); // (#14691)
-            }
-
-            storage.forEach(function (genInput: any): void {
-                if (genInput[0] === true) {
-                    createElement(
-                        SPAN, {
-                            className: PREFIX + 'annotation-title'
-                        },
-                        void 0,
-                        genInput[2]
-                    ).appendChild(doc.createTextNode(
-                        genInput[1]
-                    ));
-
-                } else {
-                    genInput[4] = {
-                        value: genInput[4][0],
-                        type: genInput[4][1]
-                    };
-                    addInput.apply(genInput[0], genInput.splice(1));
-                }
-            });
-        }
-    }
-};
-
 
 const indexFilter = /\d/g,
     PREFIX = 'highcharts-',
@@ -413,7 +114,6 @@ class Popup {
         iconsURL: string,
         chart?: Chart
     ) {
-        this.annotations = annotations;
         this.chart = chart;
         this.iconsURL = iconsURL;
         this.lang = (getOptions().lang.navigation as any).popup;
@@ -455,7 +155,6 @@ class Popup {
      *
      * */
 
-    public readonly annotations: Highcharts.PopupAnnotationsObject;
     public chart?: Chart;
     public container: HTMLDOMElement;
     public formType?: string;
@@ -847,11 +546,13 @@ class Popup {
  * */
 
 interface Popup {
-    indicators: typeof PopupIndicators;
-    tabs: typeof PopupTabs;
+    readonly annotations: typeof PopupAnnotations;
+    readonly indicators: typeof PopupIndicators;
+    readonly tabs: typeof PopupTabs;
 }
 
 extend(Popup.prototype, {
+    annotations: PopupAnnotations,
     indicators: PopupIndicators,
     tabs: PopupTabs
 });
